+++ src/apis/security/models.ts
@@ -1,0 +1,6 @@
+export interface TokenBody {
+  idToken: string;
+  accessToken: string;
+  refreshToken: string;
+  tokenType: string;
+}