+++ src/common/config.ts
@@ -1,7 +1,0 @@
-export interface ClientConfig {
-  host: string;
-}
-
-export interface Configurable {
-  config: ClientConfig;
-}