+++ src/common/index.ts
@@ -1,4 +1,4 @@
-export * from "./config";
+export * from "./client";
 // No need to import the axios config helper
 export { PortfolioRequest } from "./request";
 export * from "./response";