+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.2.0
+
+- Added security client
+  - Login functionality
+  - Confirm account functionality
+- Restructured client architecture. All clients now fall under a portfolio client.
+
 # 0.1.2
 
 - Fixed bug that incorrectly configured default api URL with extra trailing slash